% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrar-DEPRECATED.R
\name{calibrar-defunct}
\alias{calibrar-defunct}
\alias{getObservedData}
\alias{getCalibrationInfo}
\alias{createObjectiveFunction}
\title{Defunct functions in package \pkg{calibrar}.}
\usage{
getObservedData(info, path, data.folder = "data", ...)

getCalibrationInfo(
  path,
  file = "calibrationInfo.csv",
  stringsAsFactors = FALSE,
  ...
)

createObjectiveFunction(
  runModel,
  info,
  observed,
  aggFn = .weighted.sum,
  aggregate = FALSE,
  ...
)
}
\description{
The functions listed below are defunct. When possible, alternative functions with similar
  functionality are also mentioned. Help pages for defunct functions are
  available at \code{help("<function>-defunct")}.
}
\section{\code{getObservedData}}{

Deprecated in v0.3, defunct in v0.9. For \code{getObservedData}, use \code{\link{calibration_data}}.
}

\section{\code{getCalibrationInfo}}{

Deprecated in v0.3, defunct in v0.9. For \code{getCalibrationInfo}, use \code{\link{calibration_setup}}.
}

\section{\code{createObjectiveFunction}}{

Deprecated in v0.3, defunct in v0.9. For \code{createObjectiveFunction}, use \code{\link{calibration_objFn}}.
}

\keyword{internal}
