% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calibraR-main.R
\docType{package}
\name{calibrar-package}
\alias{calibrar}
\alias{calibrar-package}
\title{Automated Calibration for Complex (Ecological) Models}
\description{
Automated Calibration for Complex (Ecological) Models
}
\examples{
calibrate(par=rep(NA, 5), fn=SphereN)
calibrate(par=rep(NA, 5), fn=SphereN, replicates=3)
calibrate(par=rep(0.5, 5), fn=SphereN, replicates=3, lower=-5, upper=5)
calibrate(par=rep(0.5, 5), fn=SphereN, replicates=3, lower=-5, upper=5, phases=c(1,1,1,2,3))
calibrate(par=rep(0.5, 5), fn=SphereN, replicates=c(1,1,4), lower=-5, upper=5, phases=c(1,1,1,2,3))
}
\author{
Ricardo Oliveros-Ramos Maintainer: Ricardo Oliveros-Ramos
<ricardo.oliveros@gmail.com>
}
\references{
calibrar: an R package for the calibration of ecological models (Oliveros-Ramos and Shin 2014)
}
\keyword{calibration}

