% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrar-main.R
\name{calibration_setup}
\alias{calibration_setup}
\title{Get information to run a calibration using the \code{calibrar} package.}
\usage{
calibration_setup(file, control = list(), ...)
}
\arguments{
\item{file}{The file with the calibration information, see details.}

\item{control}{Control arguments for generating the setup. See details.}

\item{\dots}{Additional arguments to \code{read.csv} function.}
}
\value{
A data.frame with the information for the calibration of a 
model, to be used with the \code{\link{calibration_objFn}} 
and \code{\link{calibration_data}}.
}
\description{
A wrapper for \code{read.csv} checking column names and data types 
for the table with the calibration information.
}
\seealso{
\code{\link{calibration_objFn}}, \code{\link{calibration_data}}.
}
\author{
Ricardo Oliveros-Ramos
}
