% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrar-fitness.R
\name{fitness}
\alias{fitness}
\title{Calcuted error measure between observed and simulated data}
\usage{
fitness(obs, sim, FUN, ...)
}
\arguments{
\item{obs}{observed data as expected by FUN.}

\item{sim}{simulated data matching 'obs'}

\item{FUN}{the error function.}

\item{...}{}
}
\value{
the value of FUN(obs, sim)
}
\description{
Calcuted error measure between observed and simulated data
}
