% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrar-splines.R
\name{gaussian_kernel}
\alias{gaussian_kernel}
\title{Calculate a discretization of the 2D Gaussian Kernel}
\usage{
gaussian_kernel(par, lower, upper, n = 10, checkSymmetry = TRUE, ...)
}
\arguments{
\item{par}{A list, including the mean and covariance matrix.}

\item{lower}{A vector, indicating the lower bound for the calculation.}

\item{upper}{A vector, indicating the upper bound for the calculation.}

\item{n}{The number of cells for each dimension, can be one or two numbers.}

\item{checkSymmetry}{TRUE by default, checks if the covariance matrix is symmetric.}

\item{...}{Additional arguments, currently not used.}
}
\value{
A list, with 'x', 'y' and 'z' components.
}
\description{
Calculate a discretization of the 2D Gaussian Kernel
}
