% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrar-DEPRECATED.R
\name{getCalibrationInfo-defunct}
\alias{getCalibrationInfo-defunct}
\title{Get information to run a calibration using the \code{calibrar} package.}
\arguments{
\item{path}{The path to look for the file.}

\item{file}{The file with the calibration information, see details.}

\item{stringsAsFactors}{To be passed to \code{read.csv}.}

\item{\dots}{Additional arguments to \code{read.csv} function.}
}
\value{
A data.frame with the information for the calibration of a 
model, to be used with the \code{\link{createObjectiveFunction}} 
and \code{\link{getObservedData}}.
}
\description{
A wrapper for \code{read.csv} checking column names and data types 
for the table with the calibration information.
}
\seealso{
\code{\link{calibrar-defunct}}
}
\author{
Ricardo Oliveros-Ramos
}
\keyword{internal}
