% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calibraR-main.R
\name{getCalibrationInfo}
\alias{getCalibrationInfo}
\title{Get information to run a calibration using the \code{calibrar} package.}
\usage{
getCalibrationInfo(path, file = "calibrationInfo.csv",
  stringsAsFactors = FALSE, ...)
}
\arguments{
\item{path}{The path to look for the file.}

\item{file}{The file with the calibration information, see details.}

\item{stringsAsFactors}{To be passed to \code{read.csv}.}

\item{\dots}{Additional arguments to \code{read.csv} function.}
}
\value{
A data.frame with the information for the calibration of a
model, to be used with the \code{\link{createObjectiveFunction}}
and \code{\link{getObservedData}}.
}
\description{
A wrapper for \code{read.csv} checking column names and data types
for the table with the calibration information.
}
\author{
Ricardo Oliveros-Ramos
}
\seealso{
\code{\link{createObjectiveFunction}}, \code{\link{getObservedData}}.
}

