% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrar-main.R
\name{optimh}
\alias{optimh}
\title{General-purpose optimization using heuristic algorithms}
\usage{
optimh(
  par,
  fn,
  gr = NULL,
  ...,
  method = c("AHR-ES", "Nelder-Mead", "SANN", "hjn", "bobyqa", "CMA-ES", "genSA", "DE",
    "soma", "genoud", "PSO", "hybridPSO", "mads", "hjk", "hjkb", "nmk", "nmkb"),
  lower = -Inf,
  upper = +Inf,
  active = NULL,
  control = list(),
  hessian = FALSE,
  parallel = FALSE
)
}
\arguments{
\item{par}{A numeric vector or list. The length of \code{par} defines the number of
parameters to be estimated (i.e., the dimension of the problem).}

\item{fn}{The objective function to be minimised. It should accept a parameter vector
(or list, depending on the wrapper) as first argument and return either a scalar value
(single-objective) or a numeric vector (multi-objective).}

\item{gr}{Function to compute the gradient of \code{fn}. Ignored by most methods, 
added for consistency with other optimization functions.}

\item{...}{Additional arguments passed to \code{fn} and \code{gr}.}

\item{method}{Optimisation method(s) to be used. Can be a single method name or a
vector of method names (e.g., one per phase). If \code{NULL}, a default is chosen based
on \code{replicates} (see Details).}

\item{lower}{Lower bounds for parameters. One value or a vector of the same length as
\code{par}. \code{NA} is treated as \code{-Inf}. Default is unconstrained.}

\item{upper}{Upper bounds for parameters. One value or a vector of the same length as
\code{par}. \code{NA} is treated as \code{Inf}. Default is unconstrained.}

\item{active}{Boolean vector of the same length as \code{par}, indicating if the
parameter is used in the optimisation (\code{TRUE}) or held at a fixed value (\code{FALSE}).}

\item{control}{A list of control options. Common options include \code{ncores},
\code{run}, \code{master}, \code{verbose}, \code{REPORT}, \code{restart.file},
\code{gradient}, and \code{gr.method}. Additional solver-specific options may be passed
through to the underlying optimiser.}

\item{hessian}{Logical. Should a numerically differentiated Hessian matrix be returned?
Currently not implemented.}

\item{parallel}{Logical. Enable parallel computation (e.g., for replicated evaluations
and/or numerical gradients) using up to \code{control$ncores} cores.}
}
\value{
A list with components:
\describe{
\item{par}{The best set of parameters found.}
\item{value}{The value of \code{fn} corresponding to \code{par}.}
\item{counts}{A two-element integer vector giving the number of calls to \code{fn} and
\code{gr} respectively. This excludes those calls needed to compute the Hessian, if
requested, and any calls to \code{fn} to compute a finite-difference approximation to the
gradient.}
\item{convergence}{An integer code. \code{0} indicates successful completion.}
\item{message}{A character string giving any additional information returned by the
optimizer, or \code{NULL}.}
\item{hessian}{Only if argument \code{hessian} is \code{TRUE}. A symmetric matrix giving an
estimate of the Hessian at the solution found. Note that this is the Hessian of the
unconstrained problem even if the box constraints are active.}
}
}
\description{
General-purpose optimization using heuristic algorithms
}
\examples{
optim2(par=rep(NA, 5), fn=sphereN)
}
\seealso{
Other optimisers: 
\code{\link{ahres}()},
\code{\link{calibrate}()},
\code{\link{optim2}()}
}
\author{
Ricardo Oliveros-Ramos
}
\concept{optimisers}
