% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrar-class.R
\name{summary.calibrar.results}
\alias{summary.calibrar.results}
\title{Summary for calibration results object}
\usage{
\method{summary}{calibrar.results}(object, ..., show_par = NULL, par.only = FALSE)
}
\arguments{
\item{object}{Object of class calibrar.results}

\item{...}{More objects of class calibrar.results as needed for comparisons.}

\item{show_par}{Vector of names of positions of the parameters to show in the summary.}

\item{par.only}{Show only parameters in the summary, used when more than one optimization results are summarized.}
}
\description{
Summary for calibration results object
}
\keyword{internal}
