% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrar-main.R
\name{ahres}
\alias{ahres}
\title{Adaptative Hierarchical Recombination Evolutionary Strategy for 
derivative-free and black-box optimization}
\usage{
ahres(
  par,
  fn,
  gr = NULL,
  ...,
  lower = -Inf,
  upper = +Inf,
  active = NULL,
  control = list(),
  hessian = FALSE,
  parallel = FALSE
)
}
\arguments{
\item{par}{A numeric vector or list. The length of the par argument defines the 
number of parameters to be estimated (i.e. the dimension of the problem).}

\item{fn}{The function to be minimized.}

\item{gr}{the gradient of \code{fn}. Ignored, added for portability with
other optimization functions.}

\item{...}{Additional parameters to be passed to \code{fn}.}

\item{lower}{Lower threshold value(s) for parameters. One value or a vector 
of the same length as par. If one value is provided, it is used for all 
parameters. \code{NA} means \code{-Inf}. By default \code{-Inf} is used (unconstrained).}

\item{upper}{Upper threshold value(s) for parameters. One value or a vector 
of the same length as par. If one value is provided, it is used for all 
parameters. \code{NA} means \code{Inf}. By default \code{Inf} is used (unconstrained).}

\item{active}{Boolean vector of the same length as par, indicating if the 
parameter is used in the optimization (TRUE) or hold at a fixed value (FALSE).}

\item{control}{Parameter for the control of the algorithm itself, see details.}

\item{hessian}{Logical. Should a numerically differentiated Hessian matrix be returned?
Currently not implemented.}

\item{parallel}{Logical. Use parallel computation of gradients?}
}
\description{
This function performs the optimization of a function using 
the Adaptative Hierarchical Recombination Evolutionary Strategy (AHR-ES, Oliveros & Shin, 2015).
}
\examples{
ahres(par=rep(1, 5), fn=sphereN)
}
\author{
Ricardo Oliveros-Ramos
}
