\name{calibraR-package}
\alias{calibraR-package}
\alias{calibraR}
\docType{package}
\title{
Calibration of Individual Based Models using Evolutionary Algorithms
}
\description{
Calibration of Individual Based Models using Evolutionary Algorithms
}
\details{
\tabular{ll}{
Package: \tab calibraR\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2014-04-13\cr
License: \tab GLP-2\cr
}
calibration()
optimEA()
}
\author{
Ricardo Oliveros-Ramos

Maintainer: Ricardo Oliveros-Ramos <ricardo.oliveros@gmail.com>
}
\references{
Oliveros-Ramos R et al. A sequential approach to calibrate ecosystem models with multiple time series data. 
}
\keyword{ calibration }
\seealso{

\code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~

}
\examples{

}
