% Generated by roxygen2 (4.0.2): do not edit by hand
\name{calibrate}
\alias{calibrate}
\title{Sequential calibration of models}
\usage{
calibrate(par, fn, ..., aggFn = NULL, phases = NULL, replicates = 1,
  lower = -Inf, upper = Inf, gr = NULL, control = list(),
  hessian = FALSE, method = "default")
}
\arguments{
\item{par}{A numeric vector. The length of the par argument defines the
number of parameters to be estimated (i.e. the dimension of the problem).}

\item{fn}{The function to be minimized.}

\item{aggFn}{A function to aggregate \code{fn} to a scalar value if the
returned value is a vector. Some optimization algorithm can explote the
additional information provided by a vectorial output from \code{fn}.}

\item{phases}{An optional vector of the same length as \code{par},
indicating the phase at which each parameter becomes active. If omitted,
default value is 1 for all parameters, performing a single optimization.}

\item{replicates}{The number of replicates for the evaluation of \code{fn}.
The default value is 1. A value greater than 1 is only useful for stochastic
functions.}

\item{lower}{Lower threshold value(s) for parameters. One value or a vector
of the same length as par. If one value is provided, it is used for all
parameters. \code{NA} means \code{-Inf}. By default \code{-Inf} is used (unconstrained).}

\item{upper}{Upper threshold value(s) for parameters. One value or a vector
of the same length as par. If one value is provided, it is used for all
parameters. \code{NA} means \code{Inf}. By default \code{Inf} is used (unconstrained).}

\item{gr}{the gradient of \code{fn}. Ignored, added for portability with
other optimization functions.}

\item{control}{Parameter for the control of the algorithm itself, see details.}

\item{hessian}{Logical. Should a numerically differentiated Hessian matrix be returned?
Currently not implemented.}

\item{method}{The optimization method to be used. Currently, the only implemented
is the 'default' method, corresponding to the AHR-ES (Oliveros & Shin, 2014).}

\item{\dots}{Additional parametrs to be passed to \code{fn}.}

\item{restart}{If restart is an object of class \code{calibrar.restart}, it continue the
calibration using the information from that object.}
}
\description{
This function performs the optimization of a function, possibly
in sequential phases of increasing complexity, and it is designed for the
calibration of a model, by minimizing the error function \code{fn} associated to it.
}
\examples{
calibrate(par=rep(1, 5), fn=SphereN)
}
\author{
Ricardo Oliveros-Ramos
}

