% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrar-fitness.R
\name{objFn}
\alias{objFn}
\alias{fitness}
\title{Calcuted error measure between observed and simulated data}
\usage{
objFn(obs, sim, FUN, ...)

fitness(obs, sim, FUN, ...)
}
\arguments{
\item{obs}{observed data as expected by FUN.}

\item{sim}{simulated data matching 'obs'}

\item{FUN}{the error function. Current accepted values area: 'norm2', 
'lnorm2', 'lnorm3', 'multinomial', 'pois', 'penalty0', 'penalty1', 'penalty2' and 'normp'.}

\item{...}{Additional arguments to FUN}
}
\value{
the value of FUN(obs, sim, ...)
}
\description{
Calcuted error measure between observed and simulated data
}
