% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrar-splines.R
\name{spline_par}
\alias{spline_par}
\title{Predict time-varying parameters using splines.}
\usage{
spline_par(par, n, periodic = FALSE)
}
\arguments{
\item{n}{Number of knots}

\item{periodic}{boolean, is the spline periodic?}

\item{x}{Values at knots}
}
\value{

}
\description{
Predict time-varying parameters using splines.
}
