% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrar-class.R
\name{summary.calibrar.results}
\alias{summary.calibrar.results}
\title{Summary for calibration results object}
\usage{
\method{summary}{calibrar.results}(object, ..., show_par = NULL)
}
\arguments{
\item{object}{an object for which a summary is desired.}

\item{...}{additional arguments affecting the summary produced.}

\item{show_par}{Vector of names of positions of the parameters to show in the summary.}
}
\description{
Summary for calibration results object
}
